src/requirements.txt
mlflow-skinny~=2.11
numpy
pandas
fastapi
uvicorn[standard]
python-multipart
pytest
httpx
